chore: format auth and lobby modules
