+++ src/features/auth/model.js
@@ -1,0 +1,17 @@
+export function createAuthModel(authPort) {
+  return {
+    async currentUser() {
+      try {
+        const { id, email, name } = await authPort.currentUser({});
+        return { id, email, name };
+      } catch {
+        return null;
+      }
+    },
+    async logout() {
+      await authPort.logout({});
+    },
+  };
+}
+
+export default createAuthModel;
